fix: 修复编译错误
- EncryptionTaskManager: 添加 VaultConfig 序列化支持
- EncryptionProgressIcon: 移除 experimental Tooltip API
- EncryptionProgressPanel: 修复 Double/Float 类型转换

+++ app/src/main/java/com/whmdg/mczj/tools/ui/encryption/EncryptionProgressPanel.kt
@@ -426,7 +426,7 @@ private fun FourColorProgressBar(
         if (totalSize <= 0) return@Canvas
 
         // 计算各段宽度
-        val minWidth = width * 0.01
+        val minWidth = (width * 0.01).toFloat()
         var completedWidth = if (completedSize > 0) {
             (completedSize.toFloat() / totalSize * width).coerceAtLeast(minWidth)
         } else 0f
